git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/new-regalloc-branch@95203 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/tree-ssa/loop-5.c
@@ -3,11 +3,11 @@
 /* { dg-do compile } */
 /* { dg-options "-O1 -fdump-tree-vars" } */
 
-void foo(long);
+void foo(int);
 
 void xxx(void)
 {
-  long iter, jter;
+  int iter, jter;
 
   for (iter = 0, jter = 2; iter < 100; iter++, jter++)
     {
